+++ api/src/middlewares/services/UserService/personalisation/personalisationBuilder/marriage/PersonalisationBuilder.ts
@@ -1,7 +1,7 @@
-import { buildUserConfirmationPersonalisation } from "./personalisationBuilder/userConfirmation";
-import { buildUserPostalConfirmationPersonalisation } from "./personalisationBuilder/userPostalConfirmation";
+import { buildUserConfirmationPersonalisation } from "./userConfirmation";
+import { buildUserPostalConfirmationPersonalisation } from "./userPostalConfirmation";
 
-export const PersonalisationBuilder = {
+export const MarriagePersonalisationBuilder = {
   /**
    * Confirmation email for the user, if the user is applying in person, detailing how to book, what to bring etc.
    */
