+++ api/src/middlewares/services/UserService/personalisation/personalisationBuilder/marriage/userConfirmation/getAdditionalPersonalisations.ts
@@ -1,5 +1,5 @@
-import { AnswersHashMap } from "../../../../../../types/AnswersHashMap";
-import { MarriageFormType } from "../../../../../../types/FormDataBody";
+import { AnswersHashMap } from "../../../../../../../types/AnswersHashMap";
+import { MarriageFormType } from "../../../../../../../types/FormDataBody";
 
 type PersonalisationFunction = (fields: AnswersHashMap) => Record<string, boolean>;
